requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-environ==0.8.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
ImageHash==4.2.1
install==1.3.4
numpy==1.21.3
Pillow==8.4.0
PyMySQL==1.0.2
pytz==2021.3
PyWavelets==1.1.1
requests==2.26.0
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
